flask/requirements.txt
Adafruit-Blinka==4.8.1
adafruit-circuitpython-neopixel==6.0.0
adafruit-circuitpython-pypixelbuf==2.0.0
Adafruit-PlatformDetect==2.10.0
Adafruit-PureIO==1.1.5
click==7.1.1
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
NamedAtomicLock==1.1.3
pyftdi==0.51.2
pyserial==3.4
pySerialTransfer==2.1.6
python-dotenv==0.12.0
pyusb==1.0.2
rpi-ws281x==4.3.1
RPi.GPIO==0.7.0
sysv-ipc==1.0.1
Werkzeug==1.0.1
